+++ squadcastv1/src/lib/config.ts
@@ -12,7 +12,11 @@ import { Params, pathToFunc } from "./url.js";
  */
 export const ServerList = [
   /**
-   * production env
+   * production EU env
+   */
+  "https://api.eu.squadcast.com",
+  /**
+   * production US env
    */
   "https://api.squadcast.com",
 ] as const;
@@ -61,8 +65,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "1.2.0",
-  genVersion: "2.845.15",
+  sdkVersion: "1.3.0",
+  genVersion: "2.865.2",
   userAgent:
-    "speakeasy-sdk/typescript 1.2.0 2.845.15 1.0.0 @solarwinds/squadcast-sdk-typescript",
+    "speakeasy-sdk/typescript 1.3.0 2.865.2 1.0.0 @solarwinds/squadcast-sdk-typescript",
 } as const;