+++ squadcastv1/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [typescript v1.2.0] squadcastv1
 ### Releases
 - [NPM v1.2.0] https://www.npmjs.com/package/@solarwinds/squadcast-sdk-typescript/v/1.2.0 - squadcastv1
+
+## 2026-03-14 00:24:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.755.0 (2.865.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.3.0] squadcastv1
+### Releases
+- [NPM v1.3.0] https://www.npmjs.com/package/@solarwinds/squadcast-sdk-typescript/v/1.3.0 - squadcastv1